Added search box (no actual functionality) and renamed component for consistency
