+++ src/app/components/volunteering-grid/volunteering-grid.component.css
@@ -1,5 +1,6 @@
 .volunteering-grid {
     margin: 0 0 30px 0;
+    margin: 10px;
     padding: 5px;
     min-height: 400px;
     border: 1px solid #F5F5F5;
