+++ src/app/components/volunteering-search/volunteering-search.component.css
@@ -1,0 +1,5 @@
+.search-container {
+    padding: 0 0 20px 0;
+    margin: 20px 0 20px 0;
+    border-bottom: 1px solid #E5E5E5;
+}
