+++ src/app/components/volunteering-search/volunteering-search.component.html
@@ -1,0 +1,9 @@
+<div class="container search-container">
+    <div class="row">
+        <div class="col-md-4 col-xs-12">
+            <input class="form-control" #searchBox
+                   (keyup)="search(searchBox.value)"
+                   placeholder="Search for a opportunity"/>
+        </div>
+    </div>
+</div>
