+++ src/app/components/volunteering/volunteering.component.html
@@ -8,6 +8,9 @@
 <div class="container">
     <div *ngIf="opportunities">
         <div class="row">
+            <volunteering-search></volunteering-search>
+        </div>
+        <div class="row">
             <volunteering-grid *ngFor="let opportunity of displayOpportunities"
                           [opportunity]="opportunity">
             </volunteering-grid>
